di: move execution::get_env() into interface directory
