+++ libs/di/include/di/execution/query/prelude.h
@@ -1,9 +1,9 @@
 #pragma once
 
+#include <di/execution/interface/get_env.h>
 #include <di/execution/query/forwarding_query.h>
 #include <di/execution/query/get_completion_scheduler.h>
 #include <di/execution/query/get_completion_signatures.h>
 #include <di/execution/query/get_delegatee_scheduler.h>
-#include <di/execution/query/get_env.h>
 #include <di/execution/query/get_forward_progress_guarantee.h>
 #include <di/execution/query/get_scheduler.h>
